feat: add function of setting/deleting likes and deleting cards

+++ src/utils/Api.js
@@ -61,20 +61,36 @@ class Api {
     }
 
 
-    addLike(id) {
-        return fetch(`${this._url}cards/${id}/likes`, {
-            method: 'PUT',
-            headers: this._headers,
-        })
-        .then(this._checkResponse);
-    }
+    // addLike(id) {
+    //     return fetch(`${this._url}cards/${id}/likes`, {
+    //         method: 'PUT',
+    //         headers: this._headers,
+    //     })
+    //     .then(this._checkResponse);
+    // }
 
-    dislike(id) {
-        return fetch(`${this._url}cards/${id}/likes`, {
-            method: 'DELETE',
-            headers: this._headers,
-        })
-        .then(this._checkResponse);
+    // dislike(id) {
+    //     return fetch(`${this._url}cards/${id}/likes`, {
+    //         method: 'DELETE',
+    //         headers: this._headers,
+    //     })
+    //     .then(this._checkResponse);
+    // }
+
+    toggleLikeStatus(id, isLiked) {
+        if(isLiked) {
+            return fetch(`${this._url}cards/${id}/likes`, {
+                method: 'PUT',
+                headers: this._headers,
+            })
+            .then(this._checkResponse);
+        } else {
+            return fetch(`${this._url}cards/${id}/likes`, {
+                method: 'DELETE',
+                headers: this._headers,
+            })
+            .then(this._checkResponse);
+        }
     }
 
     editAvatar(data) {
